Создан абстрактный класс Component. Наследованы компоненты обычной и расширенной точек маршрута
